refactor: rule_utils -> rule
